done updateing and refactoring to L specs. Still need to work on junits
