--- src/main/java/at/refugeescode/rcstore/controller/logic/AddItemService.java
@@ -1,12 +1,0 @@
-package at.refugeescode.rcstore.controller.logic;
-
-import at.refugeescode.rcstore.models.Item;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-public interface AddItemService {
-
-    void add(MultipartFile image, Item newItem) throws IOException;
-
-}
