--- src/main/java/at/refugeescode/rcstore/controller/logic/BookingService.java
@@ -1,9 +1,0 @@
-package at.refugeescode.rcstore.controller.logic;
-
-import at.refugeescode.rcstore.models.Item;
-
-public interface BookingService {
-
-    void book(Item item);
-
-}
