--- src/main/java/at/refugeescode/rcstore/controller/logic/MyItemService.java
@@ -1,13 +1,0 @@
-package at.refugeescode.rcstore.controller.logic;
-
-import at.refugeescode.rcstore.models.Item;
-
-import java.util.List;
-
-public interface MyItemService {
-
-    List<Item> getMyItems();
-
-    void returnItem(String id);
-
-}
