--- src/main/java/at/refugeescode/rcstore/controller/logic/RegisterService.java
@@ -1,9 +1,0 @@
-package at.refugeescode.rcstore.controller.logic;
-
-import at.refugeescode.rcstore.models.UserDto;
-
-public interface RegisterService {
-
-    String controlUser(UserDto newUser);
-
-}
