+++ src/main/java/at/refugeescode/rcstore/persistence/model/UserDto.java
@@ -1,4 +1,4 @@
-package at.refugeescode.rcstore.models;
+package at.refugeescode.rcstore.persistence.model;
 
 import lombok.Data;
