--- src/main/java/at/refugeescode/rcstore/security/LoggedInUserUtility.java
@@ -1,17 +1,0 @@
-package at.refugeescode.rcstore.security;
-
-import at.refugeescode.rcstore.models.User;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class LoggedInUserUtility {
-
-    public User getLoggedOnUser() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
-        return userPrincipal.getUser();
-    }
-
-}
