+++ src/main/java/at/refugeescode/rcstore/security/UserDetailsServiceImp.java
@@ -1,6 +1,6 @@
 package at.refugeescode.rcstore.security;
 
-import at.refugeescode.rcstore.models.User;
+import at.refugeescode.rcstore.persistence.model.User;
 import at.refugeescode.rcstore.persistence.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
