+++ src/main/java/at/refugeescode/rcstore/view/controller/AddItemController.java
@@ -1,7 +1,8 @@
-package at.refugeescode.rcstore.controller.view;
+package at.refugeescode.rcstore.view.controller;
 
-import at.refugeescode.rcstore.controller.logic.AddItemService;
-import at.refugeescode.rcstore.models.Item;
+import at.refugeescode.rcstore.persistence.model.Item;
+import at.refugeescode.rcstore.view.logic.ItemService;
+import at.refugeescode.rcstore.view.logic.UsersService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -15,7 +16,8 @@ import java.io.IOException;
 @RequestMapping("/additem")
 public class AddItemController {
 
-    private final AddItemService addItemService;
+    private final ItemService itemService;
+    private final UsersService usersService;
 
     @GetMapping
     @RolesAllowed("ROLE_ADMIN")
@@ -26,14 +28,19 @@ public class AddItemController {
     @PostMapping
     @RolesAllowed("ROLE_ADMIN")
     public String addItem(@RequestParam("image") MultipartFile image, Item newItem) throws IOException {
-        addItemService.add(image, newItem);
-        return "additem";
+        itemService.add(image, newItem);
+        return "redirect:/additem";
     }
 
     @ModelAttribute("newItem")
     @RolesAllowed("ROLE_ADMIN")
     public Item getNewItem() {
         return new Item();
+    }
+
+    @ModelAttribute("admin")
+    public boolean isUserAdmin() {
+        return usersService.isLoggedOnUserAdmin();
     }
 
 }
