+++ src/main/resources/templates/home.html
@@ -23,7 +23,7 @@
         <td th:text="${item.description}">Item description</td>
         <td th:text="${item.borrowingLimit}">Item borrowing limit</td>
         <td>
-            <form th:action="@{/iteminfo/{id} (id=${item.id})}" method="GET">
+            <form th:action="@{/item/{id} (id=${item.id})}" method="GET">
                 <input type="submit" id="view" value="View" class="btn btn-primary"/>
             </form>
         </td>
