+++ src/main/resources/templates/item.html
@@ -13,12 +13,12 @@
     <p>Name: <span th:text="${item.name}">Item name</span>.</p>
     <p>Description: <span th:text="${item.description}">Item description</span>.</p>
     <p>Borrowing limit: <span th:text="${item.borrowingLimit}">Item borrowing limit</span>.</p>
-    <img th:if="${item.imageID}" th:src="@{/iteminfo/images/{id} (id=${item.imageID})}"/>
+    <img th:if="${item.imageID}" th:src="@{/item/images/{id} (id=${item.imageID})}"/>
     <div th:if="${item.borrowed}">
         <p>Borrower E-mail: <span th:text="${item.bookedBy}">Item booked by</span>.</p>
         <p>Borrowing date: <span th:text="${item.borrowingDate}">Item borrowing date</span>.</p>
     </div>
-    <form th:unless="${item.borrowed}" th:action="@{/bookitem}" th:object="${item}" method="post">
+    <form th:unless="${item.borrowed}" th:action="@{/item}" th:object="${item}" method="post">
         <input type="hidden" th:value="${item.id}" name="id"/>
         <input type="hidden" th:value="${item.name}" name="name"/>
         <input type="hidden" th:value="${item.description}" name="description"/>
