+++ src/test/java/at/refugeescode/rcstore/configuration/AdminAccountCreatorTest.java
@@ -1,6 +1,6 @@
 package at.refugeescode.rcstore.configuration;
 
-import at.refugeescode.rcstore.models.User;
+import at.refugeescode.rcstore.persistence.model.User;
 import at.refugeescode.rcstore.persistence.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
